Rename Get protocol as ReqResp
