+++ typed-transitions/src/Protocol/ReqResp/Client.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE GADTs #-}
 
-module Protocol.Get.Client where
+module Protocol.ReqResp.Client where
 
 import Protocol.Core
-import Protocol.Get.Type
+import Protocol.ReqResp.Type
 
 -- | Client reqeust with a handle for a response.
 --
@@ -13,12 +13,12 @@ data Client m request response a where
             -> (response -> m a)
             -> Client m request response a
 
--- | Interpret @'Client'@ as a client side of the typed @'GetProtocol'@
+-- | Interpret @'Client'@ as a client side of the typed @'ReqRespProtocol'@
 --
 streamClient
   :: Monad m
   => Client m request response a
-  -> Peer (GetProtocol request response) (GetMessage request response)
+  -> Peer (ReqRespProtocol request response) (ReqRespMessage request response)
           (Yielding StIdle) (Finished StDone)
           m a
 streamClient (Request request handleResponse) =
